requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.4
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
httpx==0.27.0
python-dateutil==2.9.0.post0
rich==13.7.1
